:hammer: Change from page router to app router
