+++ src/app/events/page.tsx
@@ -1,8 +1,8 @@
-import { PageHeader } from '../../components/PageHeader'
-import { BreadCrumbs } from '../../components/share/BreadCrumbs'
-import { DetailTitle } from '../../components/share/DetailTitle'
-import { EventsList } from '../../components/events/EventsList'
-import { PageFooter } from '../../components/PageFooter'
+import { PageHeader } from '@/components/PageHeader'
+import { BreadCrumbs } from '@/components/share/BreadCrumbs'
+import { DetailTitle } from '@/components/share/DetailTitle'
+import { EventsList } from '@/components/events/EventsList'
+import { PageFooter } from '@/components/PageFooter'
 
 const pages = [{ name: 'Events', href: '/events', current: true }]
