--- src/pages/_app.tsx
@@ -1,8 +1,0 @@
-import '../styles/globals.scss'
-import { AppProps } from 'next/app'
-
-function MyApp ({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
